Use last, rather than first n rows.
This feels more "natural": When limiting the stats results to n lines,
print the _last_ n lines rather than the _first_ n lines. Thus, when
sorting folders (IMAP mailboxes) by, say, largest messages, and
limiting to n folders, only the n folders with the largest messages
will be shown in the results. Use the `--reverse`switch to obtain
the n folders with the smalles messages.